refactor(configs): drop index file pattern from configs
