+++ src/commands/branch-commands/next.ts
@@ -3,7 +3,7 @@ import {
   switchBranchAction,
   TraversalDirection,
 } from '../../actions/branch_traversal';
-import { execStateConfig } from '../../lib/config';
+import { execStateConfig } from '../../lib/config/exec_state_config';
 import { profile } from '../../lib/telemetry';
 
 const args = {
