+++ src/lib/config/cache.ts
@@ -60,6 +60,4 @@ class Cache {
   }
 }
 
-const cache = new Cache();
-
-export default cache;
+export const cache = new Cache();
