--- src/lib/config/index.ts
@@ -1,15 +1,0 @@
-import cache from './cache';
-import execStateConfig from './exec_state_config';
-import {
-  getOwnerAndNameFromURLForTesting,
-  repoConfigFactory,
-} from './repo_config';
-import { getRepoRootPath } from './repo_root_path';
-
-export {
-  repoConfigFactory,
-  getOwnerAndNameFromURLForTesting,
-  execStateConfig,
-  cache,
-  getRepoRootPath,
-};
